requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
sentence-transformers==3.1.1
numpy==1.26.4
python-dotenv==1.0.1